fix(mitre): build the case MITRE table from event tags, not only from AI synthesis
The dashboard MITRE ATT&CK panel rendered a single em-dash on a case holding nine
techniques, in all four profiles that include it — worst in the Report profile,
where the technique list is what a reader lifts into a deliverable. The report's own
MITRE section reads the same aggregate and was equally empty.

Cause: every deterministic importer emits its delta with a hardcoded top-level
`mitreTechniques: []` and carries the real technique ids on `forensicEvents[]`
instead. mergeDelta only ever collected the top-level array, so the aggregate stayed
empty until an AI synthesis pass happened to union the per-event tags in. Kill Chain
and the ATT&CK Navigator export read the events directly, which is why those two
surfaces looked correct and made this read as a rendering bug.

mergeDelta now unions the techniques carried by the merged timeline into the
aggregate, so any delta contributes and an existing case heals on its next merge.
unionEventTechniques is already pure and idempotent, so re-merging the same events
changes nothing.

That import crossed a layer: stateMerge is analysis/timeline (tier 0) and
reconTechniques is analysis/intel (tier 1). Rather than record a new boundary
violation, the dependency-free half of that module — the ATT&CK id -> name table,
techniqueName and unionEventTechniques — moves to analysis/attackTechniqueNames.ts
at tier 0, which both callers may reach. Matching command lines to techniques is
intel work and stays in reconTechniques; naming a technique is reference data.
reconTechniques re-exports both so existing importers are unchanged.

Closes #878
