+++ companion/scripts/module-map.json
@@ -160,6 +160,7 @@
     "attack.ts": "analysis/intel",
     "attackMitigations.ts": "analysis/intel",
     "attackMitigationsData.ts": "analysis/intel",
+    "attackTechniqueNames.ts": "analysis/timeline",
     "auditdImport.ts": "analysis/ingest",
     "awsImport.ts": "analysis/ingest",
     "bashHistoryImport.ts": "analysis/ingest",
